Added Python backend
This was just the testing prior to this project that allowed me to understand the basis and mathematics behind the backend graphics rendering itself. At the time I was not familiar with C++, and so used Python in order to better understand what was going on. In time this will hopefully be enhanced and converted into C++ in order to actually work with the rest of the project.